restructured and refactored
