+++ rest-service/src/main/java/no/group7/restservice/entity/Poll.java
@@ -1,4 +1,4 @@
-package com.example.restservice;
+package no.group7.restservice.entity;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
