+++ rest-service/src/main/java/no/group7/restservice/entity/Vote.java
@@ -1,4 +1,4 @@
-package com.example.restservice;
+package no.group7.restservice.entity;
 
 import javax.persistence.*;
 import java.util.Objects;
