+++ rest-service/src/main/java/no/group7/restservice/exception/FieldNotFound.java
@@ -1,4 +1,4 @@
-package com.example.restservice.exception;
+package no.group7.restservice.exception;
 
 public class FieldNotFound extends RuntimeException {
