+++ rest-service/src/main/java/no/group7/restservice/repository/PollRepository.java
@@ -1,6 +1,6 @@
-package com.example.restservice.repository;
+package no.group7.restservice.repository;
 
-import com.example.restservice.Poll;
+import no.group7.restservice.entity.Poll;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PollRepository extends JpaRepository<Poll, Long> {
