+++ rest-service/src/main/java/no/group7/restservice/repository/VoteRepository.java
@@ -1,6 +1,6 @@
-package com.example.restservice.repository;
+package no.group7.restservice.repository;
 
-import com.example.restservice.Vote;
+import no.group7.restservice.entity.Vote;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface VoteRepository extends JpaRepository<Vote, Long> {
